requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
boto3==1.40.39
botocore==1.40.39
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iso3166==2.1.1
jmespath==1.0.1
Mako==1.3.10
mapbox==0.18.1
MarkupSafe==3.0.2
msgpack==1.1.1
passlib==1.7.4
polyline==2.0.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
supabase
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
